requirements.txt
python==3.10.0
tensorflow==2.10.0
tensorflow-gpu==2.10.0
opencv-python
imageio
pillow
scikit-learn